+++ travis_emulator/travis_emulator/scripts/main.py
@@ -45,7 +45,7 @@ import gzip
 import shutil
 
 
-__version__ = "2.0.4"
+__version__ = "2.0.5"
 
 
 def fake_setup(**kwargs):