+++ z0bug_odoo/z0bug_odoo/__init__.py
@@ -20,7 +20,7 @@ except ImportError:
     except ImportError:
         release = ''
 
-__version__ = '2.0.7'
+__version__ = '2.0.8'
 if eval(os.environ.get('TRAVIS_DEBUG_MODE', '0')) > 2:
     print('DEBUG: z0bug_odoo %s' % __version__)
     print('DEBUG: z0bug_odoo.sys.path=%s' % sys.path)